refactor(photos): move photos out of the unused hd wallet package
also adds some more logging

closes #33
