--- repo/images/images.go
@@ -1,43 +1,0 @@
-package images
-
-import (
-	"encoding/base64"
-	"github.com/nfnt/resize"
-	"image"
-	_ "image/gif"
-	_ "image/png"
-	"strings"
-)
-
-func DecodeImageData(base64ImageData string) (image.Image, *image.Config, error) {
-	reader := base64.NewDecoder(base64.StdEncoding, strings.NewReader(base64ImageData))
-	img, _, err := image.Decode(reader)
-	if err != nil {
-		return nil, nil, err
-	}
-	reader = base64.NewDecoder(base64.StdEncoding, strings.NewReader(base64ImageData))
-	imgCfg, _, err := image.DecodeConfig(reader)
-	if err != nil {
-		return nil, nil, err
-	}
-	return img, &imgCfg, err
-}
-
-func ResizeImage(img image.Image, imgCfg *image.Config, w, h uint) image.Image {
-	width, height := getImageAttributes(w, h, uint(imgCfg.Width), uint(imgCfg.Height))
-	return resize.Resize(width, height, img, resize.Lanczos3)
-}
-
-func getImageAttributes(targetWidth, targetHeight, imgWidth, imgHeight uint) (width, height uint) {
-	targetRatio := float32(targetWidth) / float32(targetHeight)
-	imageRatio := float32(imgWidth) / float32(imgHeight)
-	var h, w float32
-	if imageRatio > targetRatio {
-		h = float32(targetHeight)
-		w = float32(targetHeight) * imageRatio
-	} else {
-		w = float32(targetWidth)
-		h = float32(targetWidth) * (float32(imgHeight) / float32(imgWidth))
-	}
-	return uint(w), uint(h)
-}
