+++ repo/photos/photos_test.go
@@ -1,0 +1,7 @@
+package photos
+
+import (
+	"testing"
+)
+
+func TestImageFormats(t *testing.T) {}
